Squashed commit of the following:
commit be7bb2e39ffdec315ef0d273566282c7c146ae6a
Merge: c252f5f 1c30728
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 17:05:33 2023 -0300

    Merge branch 'release/1.4.0'

commit 1c30728880874eac25f7373a1f8c702042b24682
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 17:05:18 2023 -0300

    version: 1.4.0

commit 8d91e7cb1d04abb2cf01e703a59222f04805fc5e
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 16:21:29 2023 -0300

    feat: Added encoding option in endpoint sendWhatsAppAudio

commit 68d980795ad3fc57a4b902945be094934d9095cd
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 15:13:46 2023 -0300

    fix: fix in chatwoot set, sign msg can now be disabled

commit 45c11a5a8e9474a580557d8a8aa50f200d34646c
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 15:12:13 2023 -0300

    fix: fix in chatwoot set, sign msg can now be disabled

commit 4d00351db790b498f919f3083787a85ad7788ecb
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 15:05:26 2023 -0300

    fix: command to create new instances set to /new_instance:<NAME>:<NUMBER>

commit fff420b652120af7919c01b9a4c55445da2bca9b
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 15:05:03 2023 -0300

    fix: command to create new instances set to /new_instance:<NAME>:<NUMBER>

commit 7103a953056929c7a5ab215cdaccaadfb8d8b528
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 13:43:18 2023 -0300

    feat: Added connection with pairing code in chatwoot

commit bcada5d553aa7c79f35fc825ea3dc9be2c4febca
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 13:03:40 2023 -0300

    fix: Now accepts all chatwoot inbox templates

commit c9b24ff612dbdcaa95cead2509fb30ed1ac5c2a6
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 12:55:27 2023 -0300

    fix: adjusts for chatwoot v3

commit 854c7ed04d3c50cae6ed31ce4e44ac41d392d0e3
Merge: 1aa837d c005495
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 12:02:38 2023 -0300

    fix: Improvement for Validation of Numbers and Groups

commit c0054959cdd2cb1029a3db9e1af08ecfe66e6575
Merge: 20abdd2 95df402
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 12:00:33 2023 -0300

    fix: Improvement for Validation of Numbers and Groups

    Improvement for Validation of Numbers and Groups

commit 1aa837d220e638c7f0a91d97fcee8f490c5d819a
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 11:58:52 2023 -0300

    fix: Adjusts in chatwoot integration

commit 2f3d6f7e63b556d5dfcd2f2c69c57fa78c4a5570
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 11:48:10 2023 -0300

    fix: Fixed problem with fileSha256 appearing when sending a sticker in chatwoot

commit a73d5f4b4dbed0d797436afff1168dc2e3c9a76d
Merge: f35b62e 20abdd2
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 11:18:12 2023 -0300

    Merge branch 'develop' of github.com:EvolutionAPI/evolution-api into develop

commit f35b62ed12fdd6a0c184f042eb800e61d8959e83
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 11:18:08 2023 -0300

    fix: Adjusts in createJid

commit 20abdd29087e866e4f008b7e23149fabbe4cc253
Merge: 3ca8ab1 28c2c72
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 11:17:35 2023 -0300

    fix: Adjusts in createJid to sendContact

    Adjusts in createJid to sendContact

commit 3ca8ab12a42d14141166b370f1186c513e514373
Merge: 1fcbd4f fd82aa1
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 11:13:30 2023 -0300

    fix: Adjusts in createJid

    Adjusts in createJid

commit 1fcbd4f9fd6f6f3024baf8bb7a3f3af7179dc56b
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 10:18:16 2023 -0300

    feat: Added reject call and send text message when receiving a call and Added setting to ignore group messages

commit 73d9cd62a56d56c070c54ccf39134b6e63f9a6c5
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Mon Jul 24 09:42:29 2023 -0300

    feat: Created settings Controller

commit be699d24a1d31b17670293564b1c937d1255fc4b
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Sun Jul 23 22:26:38 2023 -0300

    feat: Added fetch profile endpoint in chat controller and link preview option in send text message

commit 798eb90bed2873be514135512b75d14869777b09
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Sun Jul 23 22:24:21 2023 -0300

    feat: Added fetch profile endpoint in chat controller and link preview option in send text message

commit c252f5f8d9d894b7a3d6531eed3da2eb19fffff9
Merge: 69f5cdd 76d77ad
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Sun Jul 23 22:05:38 2023 -0300

    Merge pull request #20 from EvolutionAPI/revert-18-fetch-profile

    Revert "Adição de FetchProfile"

commit 76d77ad76f54283fe47aa0376b89897055f73220
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Sun Jul 23 22:05:21 2023 -0300

    Revert "Adição de FetchProfile"

commit 69f5cdd61a0d24dd34ab98c30c8377c44518c111
Merge: 40c230c 1ec3ed3
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Sun Jul 23 21:30:11 2023 -0300

    feat: added fetchProfile endpoint in chat controller

    Added fetchProfile endpoint in chat controller

commit 9f52f20660223b8ba3d5594ccd38e60bcdcd4e8b
Merge: 16ed582 90048af
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Sun Jul 23 21:25:20 2023 -0300

    feat: Add LinkPreview Option

    Add LinkPreview Option

commit 90048afa9d1bb700380c3b28497241a8a386607a
Author: Alan Mosko <moskoweb@gmail.com>
Date:   Sun Jul 23 11:24:16 2023 -0300

    Add LinkPreview Option

commit 16ed5821e277a4128026bcf8137448dfbc5db9da
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Fri Jul 21 20:58:37 2023 -0300

    Added connection functionality via pairing code

commit b681e339444859ba94f6d3d9432642873bb54a07
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Fri Jul 21 20:38:10 2023 -0300

    Added connection functionality via pairing code

commit 8f4d44a21212cf3d075b93c3cb2fae5a6acdaf88
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Fri Jul 21 20:37:58 2023 -0300

    Added connection functionality via pairing code

commit 93a5d07f9a9d7c432f24c73cdc72c09a9bd214ef
Merge: d0fa3b9 40c230c
Author: Davidson Gomes <davidsongviolao@gmail.com>
Date:   Fri Jul 21 17:20:19 2023 -0300

    Merge tag '1.3.2' into develop

    * Fix in update settings that needed to restart after updated
    * Correction in the use of the api with mongodb
    * Adjustments to search endpoint for contacts, chats, messages and Status messages
    * Now when deleting the instance, the data referring to it in mongodb is also deleted
    * It is now validated if the instance name contains uppercase and special characters
    * For compatibility reasons, container mode has been removed
    * Added docker-compose files example

+++ src/whatsapp/services/whatsapp.service.ts
@@ -84,8 +84,8 @@ import { arrayUnique, isBase64, isURL } from 'class-validator';
 import {
   ArchiveChatDto,
   DeleteMessage,
-  OnWhatsAppDto,
   NumberBusiness,
+  OnWhatsAppDto,
   PrivacySettingDto,
   ReadMessageDto,
   WhatsAppNumberDto,
@@ -125,6 +125,7 @@ import { Log } from '../../config/env.config';
 import ProxyAgent from 'proxy-agent';
 import { ChatwootService } from './chatwoot.service';
 import { waMonitor } from '../whatsapp.module';
+import { SettingsRaw } from '../models';
 
 export class WAStartupService {
   constructor(
@@ -143,12 +144,15 @@ export class WAStartupService {
   public client: WASocket;
   private readonly localWebhook: wa.LocalWebHook = {};
   private readonly localChatwoot: wa.LocalChatwoot = {};
+  private readonly localSettings: wa.LocalSettings = {};
   private stateConnection: wa.StateConnection = { state: 'close' };
   public readonly storePath = join(ROOT_DIR, 'store');
   private readonly msgRetryCounterCache: CacheStore = new NodeCache();
   private readonly userDevicesCache: CacheStore = new NodeCache();
   private endSession = false;
   private logBaileys = this.configService.get<Log>('LOG').BAILEYS;
+
+  private phoneNumber: string;
 
   private chatwootService = new ChatwootService(waMonitor, this.configService);
 
@@ -239,6 +243,12 @@ export class WAStartupService {
 
   public get qrCode(): wa.QrCode {
     this.logger.verbose('Getting qrcode');
+    if (this.instance.qrcode?.pairingCode) {
+      return {
+        pairingCode: this.instance.qrcode?.pairingCode,
+      };
+    }
+
     return {
       code: this.instance.qrcode?.code,
       base64: this.instance.qrcode?.base64,
@@ -338,6 +348,46 @@ export class WAStartupService {
     this.logger.verbose(`Chatwoot inbox name: ${data.name_inbox}`);
     this.logger.verbose(`Chatwoot sign msg: ${data.sign_msg}`);
 
+    return data;
+  }
+
+  private async loadSettings() {
+    this.logger.verbose('Loading settings');
+    const data = await this.repository.settings.find(this.instanceName);
+    this.localSettings.reject_call = data?.reject_call;
+    this.logger.verbose(`Settings reject_call: ${this.localSettings.reject_call}`);
+
+    this.localSettings.msg_call = data?.msg_call;
+    this.logger.verbose(`Settings msg_call: ${this.localSettings.msg_call}`);
+
+    this.localSettings.groups_ignore = data?.groups_ignore;
+    this.logger.verbose(`Settings groups_ignore: ${this.localSettings.groups_ignore}`);
+
+    this.logger.verbose('Settings loaded');
+  }
+
+  public async setSettings(data: SettingsRaw) {
+    this.logger.verbose('Setting settings');
+    await this.repository.settings.create(data, this.instanceName);
+    this.logger.verbose(`Settings reject_call: ${data.reject_call}`);
+    this.logger.verbose(`Settings msg_call: ${data.msg_call}`);
+    this.logger.verbose(`Settings groups_ignore: ${data.groups_ignore}`);
+    Object.assign(this.localSettings, data);
+    this.logger.verbose('Settings set');
+  }
+
+  public async findSettings() {
+    this.logger.verbose('Finding settings');
+    const data = await this.repository.settings.find(this.instanceName);
+
+    if (!data) {
+      this.logger.verbose('Settings not found');
+      throw new NotFoundException('Settings not found');
+    }
+
+    this.logger.verbose(`Settings url: ${data.reject_call}`);
+    this.logger.verbose(`Settings msg_call: ${data.msg_call}`);
+    this.logger.verbose(`Settings groups_ignore: ${data.groups_ignore}`);
     return data;
   }
 
@@ -556,6 +606,13 @@ export class WAStartupService {
         color: { light: '#ffffff', dark: '#198754' },
       };
 
+      if (this.phoneNumber) {
+        await delay(2000);
+        this.instance.qrcode.pairingCode = await this.client.requestPairingCode(
+          this.phoneNumber,
+        );
+      }
+
       this.logger.verbose('Generating QR code');
       qrcode.toDataURL(qr, optsQrcode, (error, base64) => {
         if (error) {
@@ -567,7 +624,12 @@ export class WAStartupService {
         this.instance.qrcode.code = qr;
 
         this.sendDataWebhook(Events.QRCODE_UPDATED, {
-          qrcode: { instance: this.instance.name, code: qr, base64 },
+          qrcode: {
+            instance: this.instance.name,
+            pairingCode: this.instance.qrcode.pairingCode,
+            code: qr,
+            base64,
+          },
         });
 
         if (this.localChatwoot.enabled) {
@@ -575,7 +637,12 @@ export class WAStartupService {
             Events.QRCODE_UPDATED,
             { instanceName: this.instance.name },
             {
-              qrcode: { instance: this.instance.name, code: qr, base64 },
+              qrcode: {
+                instance: this.instance.name,
+                pairingCode: this.instance.qrcode.pairingCode,
+                code: qr,
+                base64,
+              },
             },
           );
         }
@@ -584,7 +651,7 @@ export class WAStartupService {
       this.logger.verbose('Generating QR code in terminal');
       qrcodeTerminal.generate(qr, { small: true }, (qrcode) =>
         this.logger.log(
-          `\n{ instance: ${this.instance.name}, qrcodeCount: ${this.instance.qrcode.count} }\n` +
+          `\n{ instance: ${this.instance.name} pairingCode: ${this.instance.qrcode.pairingCode}, qrcodeCount: ${this.instance.qrcode.count} }\n` +
             qrcode,
         ),
       );
@@ -751,11 +818,12 @@ export class WAStartupService {
     return await useMultiFileAuthState(join(INSTANCE_DIR, this.instance.name));
   }
 
-  public async connectToWhatsapp(): Promise<WASocket> {
+  public async connectToWhatsapp(number?: string): Promise<WASocket> {
     this.logger.verbose('Connecting to whatsapp');
     try {
       this.loadWebhook();
       this.loadChatwoot();
+      this.loadSettings();
 
       this.instance.authState = await this.defineAuthState();
 
@@ -823,6 +891,15 @@ export class WAStartupService {
       this.eventHandler();
 
       this.logger.verbose('Socket event handler initialized');
+
+      this.phoneNumber = number;
+
+      // if (number) {
+      //   this.logger.verbose('creating pairing code');
+      //   await delay(5000);
+      //   this.phoneNumber = number;
+      //   this.instance.qrcode.pairingCode = await this.client.requestPairingCode(number);
+      // }
 
       return this.client;
     } catch (error) {
@@ -1038,13 +1115,14 @@ export class WAStartupService {
         type: MessageUpsertType;
       },
       database: Database,
+      settings: SettingsRaw,
     ) => {
       this.logger.verbose('Event received: messages.upsert');
       const received = messages[0];
 
       if (
         type !== 'notify' ||
-        received.message?.protocolMessage ||
+        // received.message?.protocolMessage ||
         received.message?.pollUpdateMessage
       ) {
         this.logger.verbose('message rejected');
@@ -1053,6 +1131,11 @@ export class WAStartupService {
 
       if (Long.isLong(received.messageTimestamp)) {
         received.messageTimestamp = received.messageTimestamp?.toNumber();
+      }
+
+      if (settings.groups_ignore && received.key.remoteJid.includes('@g.us')) {
+        this.logger.verbose('group ignored');
+        return;
       }
 
       const messageRaw: MessageRaw = {
@@ -1146,7 +1229,11 @@ export class WAStartupService {
       );
     },
 
-    'messages.update': async (args: WAMessageUpdate[], database: Database) => {
+    'messages.update': async (
+      args: WAMessageUpdate[],
+      database: Database,
+      settings: SettingsRaw,
+    ) => {
       this.logger.verbose('Event received: messages.update');
       const status: Record<number, wa.StatusMessage> = {
         0: 'ERROR',
@@ -1157,6 +1244,10 @@ export class WAStartupService {
         5: 'PLAYED',
       };
       for await (const { key, update } of args) {
+        if (settings.groups_ignore && key.remoteJid.includes('@g.us')) {
+          this.logger.verbose('group ignored');
+          return;
+        }
         if (key.remoteJid !== 'status@broadcast' && !key?.remoteJid?.match(/(:\d+)/)) {
           this.logger.verbose('Message update is valid');
 
@@ -1256,9 +1347,32 @@ export class WAStartupService {
 
   private eventHandler() {
     this.logger.verbose('Initializing event handler');
-    this.client.ev.process((events) => {
+    this.client.ev.process(async (events) => {
       if (!this.endSession) {
         const database = this.configService.get<Database>('DATABASE');
+        const settings = await this.findSettings();
+
+        if (events.call) {
+          this.logger.verbose('Listening event: call');
+          const call = events.call[0];
+
+          if (settings?.reject_call && call.status == 'offer') {
+            this.logger.verbose('Rejecting call');
+            this.client.rejectCall(call.id, call.from);
+          }
+
+          if (settings?.msg_call.trim().length > 0 && call.status == 'offer') {
+            this.logger.verbose('Sending message in call');
+            const msg = await this.client.sendMessage(call.from, {
+              text: settings.msg_call,
+            });
+
+            this.client.ev.emit('messages.upsert', {
+              messages: [msg],
+              type: 'notify',
+            });
+          }
+        }
 
         if (events['connection.update']) {
           this.logger.verbose('Listening event: connection.update');
@@ -1279,37 +1393,44 @@ export class WAStartupService {
         if (events['messages.upsert']) {
           this.logger.verbose('Listening event: messages.upsert');
           const payload = events['messages.upsert'];
-          this.messageHandle['messages.upsert'](payload, database);
+          this.messageHandle['messages.upsert'](payload, database, settings);
         }
 
         if (events['messages.update']) {
           this.logger.verbose('Listening event: messages.update');
           const payload = events['messages.update'];
-          this.messageHandle['messages.update'](payload, database);
+          this.messageHandle['messages.update'](payload, database, settings);
         }
 
         if (events['presence.update']) {
           this.logger.verbose('Listening event: presence.update');
           const payload = events['presence.update'];
+
+          if (settings.groups_ignore && payload.id.includes('@g.us')) {
+            this.logger.verbose('group ignored');
+            return;
+          }
           this.sendDataWebhook(Events.PRESENCE_UPDATE, payload);
         }
 
-        if (events['groups.upsert']) {
-          this.logger.verbose('Listening event: groups.upsert');
-          const payload = events['groups.upsert'];
-          this.groupHandler['groups.upsert'](payload);
-        }
+        if (!settings?.groups_ignore) {
+          if (events['groups.upsert']) {
+            this.logger.verbose('Listening event: groups.upsert');
+            const payload = events['groups.upsert'];
+            this.groupHandler['groups.upsert'](payload);
+          }
 
-        if (events['groups.update']) {
-          this.logger.verbose('Listening event: groups.update');
-          const payload = events['groups.update'];
-          this.groupHandler['groups.update'](payload);
-        }
+          if (events['groups.update']) {
+            this.logger.verbose('Listening event: groups.update');
+            const payload = events['groups.update'];
+            this.groupHandler['groups.update'](payload);
+          }
 
-        if (events['group-participants.update']) {
-          this.logger.verbose('Listening event: group-participants.update');
-          const payload = events['group-participants.update'];
-          this.groupHandler['group-participants.update'](payload);
+          if (events['group-participants.update']) {
+            this.logger.verbose('Listening event: group-participants.update');
+            const payload = events['group-participants.update'];
+            this.groupHandler['group-participants.update'](payload);
+          }
         }
 
         if (events['chats.upsert']) {
@@ -1381,7 +1502,7 @@ export class WAStartupService {
 
   private createJid(number: string): string {
     this.logger.verbose('Creating jid with number: ' + number);
-    
+
     if (number.includes('@g.us') || number.includes('@s.whatsapp.net')) {
       this.logger.verbose('Number already contains @g.us or @s.whatsapp.net');
       return number;
@@ -1391,7 +1512,7 @@ export class WAStartupService {
       this.logger.verbose('Number already contains @broadcast');
       return number;
     }
-    
+
     number = number
       ?.replace(/\s/g, '')
       .replace(/\+/g, '')
@@ -1399,21 +1520,15 @@ export class WAStartupService {
       .replace(/\)/g, '')
       .split(/\:/)[0]
       .split('@')[0];
-    
-    if(number.includes('-') && number.length >= 24){
+
+    if (number.includes('-') && number.length >= 24) {
       this.logger.verbose('Jid created is group: ' + `${number}@g.us`);
       number = number.replace(/[^\d-]/g, '');
       return `${number}@g.us`;
     }
-    
+
     number = number.replace(/\D/g, '');
-    
-    if (number.length >= 18) {
-      this.logger.verbose('Jid created is group: ' + `${number}@g.us`);
-      number = number.replace(/[^\d-]/g, '');
-      return `${number}@g.us`;
-    }
-    
+
     this.logger.verbose('Jid created is whatsapp: ' + `${number}@s.whatsapp.net`);
     return `${number}@s.whatsapp.net`;
   }
@@ -1436,10 +1551,10 @@ export class WAStartupService {
       };
     }
   }
-  
+
   public async getStatus(number: string) {
     const jid = this.createJid(number);
-    
+
     this.logger.verbose('Getting profile status with jid:' + jid);
     try {
       this.logger.verbose('Getting status');
@@ -1455,21 +1570,21 @@ export class WAStartupService {
       };
     }
   }
-  
+
   public async fetchProfile(instanceName: string, number?: string) {
-    const jid = (number)
-      ? this.createJid(number)
-      : this.client?.user?.id;
+    const jid = number ? this.createJid(number) : this.client?.user?.id;
+
     this.logger.verbose('Getting profile with jid: ' + jid);
     try {
       this.logger.verbose('Getting profile info');
+      const info = await waMonitor.instanceInfo(instanceName);
       const business = await this.fetchBusinessProfile(jid);
-      
+
       if (number) {
         const info = (await this.whatsappNumber({ numbers: [jid] }))?.shift();
         const picture = await this.profilePicture(jid);
         const status = await this.getStatus(jid);
-        
+
         return {
           wuid: jid,
           name: info?.name,
@@ -1483,7 +1598,7 @@ export class WAStartupService {
         };
       } else {
         const info = await waMonitor.instanceInfo(instanceName);
-        
+
         return {
           wuid: jid,
           name: info?.instance?.profileName,
@@ -1496,7 +1611,6 @@ export class WAStartupService {
           website: business?.website?.shift(),
         };
       }
-      
     } catch (error) {
       this.logger.verbose('Profile not found');
       return {
@@ -1547,6 +1661,8 @@ export class WAStartupService {
       
       // Link Preview
       let linkPreview = (options?.linkPreview != false) ? undefined : false;
+
+      const linkPreview = options?.linkPreview != false ? undefined : false;
 
       let quoted: WAMessage;
 
@@ -2063,26 +2179,45 @@ export class WAStartupService {
 
   public async audioWhatsapp(data: SendAudioDto) {
     this.logger.verbose('Sending audio whatsapp');
-    const convert = await this.processAudio(data.audioMessage.audio, data.number);
-    if (typeof convert === 'string') {
-      const audio = fs.readFileSync(convert).toString('base64');
-      const result = this.sendMessageWithTyping<AnyMessageContent>(
-        data.number,
-        {
-          audio: Buffer.from(audio, 'base64'),
-          ptt: true,
-          mimetype: 'audio/mp4',
-        },
-        { presence: 'recording', delay: data?.options?.delay },
-      );
 
-      fs.unlinkSync(convert);
-      this.logger.verbose('Converted audio deleted');
-
-      return result;
-    } else {
-      throw new InternalServerErrorException(convert);
+    if (!data.options?.encoding && data.options?.encoding !== false) {
+      data.options.encoding = true;
     }
+
+    if (data.options?.encoding) {
+      const convert = await this.processAudio(data.audioMessage.audio, data.number);
+      if (typeof convert === 'string') {
+        const audio = fs.readFileSync(convert).toString('base64');
+        const result = this.sendMessageWithTyping<AnyMessageContent>(
+          data.number,
+          {
+            audio: Buffer.from(audio, 'base64'),
+            ptt: true,
+            mimetype: 'audio/mp4',
+          },
+          { presence: 'recording', delay: data?.options?.delay },
+        );
+
+        fs.unlinkSync(convert);
+        this.logger.verbose('Converted audio deleted');
+
+        return result;
+      } else {
+        throw new InternalServerErrorException(convert);
+      }
+    }
+
+    return await this.sendMessageWithTyping<AnyMessageContent>(
+      data.number,
+      {
+        audio: isURL(data.audioMessage.audio)
+          ? { url: data.audioMessage.audio }
+          : Buffer.from(data.audioMessage.audio, 'base64'),
+        ptt: true,
+        mimetype: 'audio/ogg; codecs=opus',
+      },
+      { presence: 'recording', delay: data?.options?.delay },
+    );
   }
 
   public async buttonMessage(data: SendButtonDto) {
@@ -2246,7 +2381,7 @@ export class WAStartupService {
     const onWhatsapp: OnWhatsAppDto[] = [];
     for await (const number of data.numbers) {
       let jid = this.createJid(number);
-      
+
       if (isJidGroup(jid)) {
         const group = await this.findGroup({ groupJid: jid }, 'inner');
 
@@ -2254,7 +2389,7 @@ export class WAStartupService {
 
         onWhatsapp.push(new OnWhatsAppDto(group.id, !!group?.id, group?.subject));
       } else {
-        jid = (!jid.startsWith('+')) ? `+${jid}` : jid;
+        jid = !jid.startsWith('+') ? `+${jid}` : jid;
         const verify = await this.client.onWhatsApp(jid);
 
         const result = verify[0];
@@ -2536,30 +2671,28 @@ export class WAStartupService {
     }
   }
 
-  public async fetchBusinessProfile(number: string) : Promise<NumberBusiness> {
+  public async fetchBusinessProfile(number: string): Promise<NumberBusiness> {
     this.logger.verbose('Fetching business profile');
     try {
-      const jid = (number) 
-        ? this.createJid(number)
-        : this.instance.wuid;
+      const jid = number ? this.createJid(number) : this.instance.wuid;
 
       const profile = await this.client.getBusinessProfile(jid);
       this.logger.verbose('Trying to get business profile');
 
       if (!profile) {
         const info = await this.whatsappNumber({ numbers: [jid] });
-        
+
         return {
           isBusiness: false,
           message: 'Not is business profile',
-          ...info?.shift()
+          ...info?.shift(),
         };
       }
 
       this.logger.verbose('Business profile fetched');
       return {
         isBusiness: true,
-        ...profile
+        ...profile,
       };
     } catch (error) {
       throw new InternalServerErrorException(
